+++ R/mod_app.R
@@ -14,14 +14,26 @@ mod_app_ui <- function(id, ide_colors = get_ide_theme_info()) {
     bslib::page_fluid(
       theme = create_chat_app_theme(ide_colors),
       title = "ChatGPT from gptstudio",
-      class = "vh-100 p-3 m-0",
+      class = "vh-100 p-0 m-0",
       html_dependencies(),
-      div(
-        class = "row justify-content-center h-100",
+
+      bslib::layout_sidebar(
+        class = "vh-100",
+        sidebar = bslib::sidebar(
+          open = "closed",
+          width = 300,
+          class = "p-0",
+          padding = "0.5rem",
+
+          mod_sidebar_ui(ns("sidebar"), translator)
+        ),
         div(
-          class = "col h-100",
-          style = htmltools::css(`max-width` = "800px"),
-          mod_chat_ui(ns("chat"), translator)
+          class = "row justify-content-center h-100",
+          div(
+            class = "col h-100",
+            style = htmltools::css(`max-width` = "800px"),
+            mod_chat_ui(ns("chat"), translator)
+          )
         )
       )
     )
@@ -35,7 +47,14 @@ mod_app_ui <- function(id, ide_colors = get_ide_theme_info()) {
 #'
 mod_app_server <- function(id, ide_colors = get_ide_theme_info()) {
   moduleServer(id, function(input, output, session) {
-    mod_chat_server("chat", ide_colors)
+    sidebar <- mod_sidebar_server("sidebar")
+    mod_chat_server(
+      id = "chat",
+      ide_colors = ide_colors,
+      translator = NULL,
+      settings = sidebar$settings,
+      history = sidebar$history
+    )
   })
 }
 
@@ -69,7 +88,7 @@ create_chat_app_theme <- function(ide_colors = get_ide_theme_info()) {
     version = 5,
     bg = ide_colors$bg,
     fg = ide_colors$fg,
-    font_scale = 0.9
+    font_scale = 0.9,
   )
 }
 
@@ -109,10 +128,10 @@ get_ide_theme_info <- function() {
 
 html_dependencies <- function() {
   htmltools::htmlDependency(
-    name = "gptstudio-assets", version = "0.2.0",
+    name = "gptstudio-assets", version = "0.4.0",
     package = "gptstudio",
     src = "assets",
-    script = c("js/copyToClipboard.js", "js/shiftEnter.js"),
+    script = c("js/copyToClipboard.js", "js/shiftEnter.js", "js/conversation.js"),
     stylesheet = c("css/mod_app.css")
   )
 }